Restructure detection tests to match other suites
- Move detection tests from tests/ to tests/detection/ (independent suite)
- Rename test.yml to test-detection.yml with updated paths
- Update all import paths in test files and utils
- Update README.md with new structure
- Update copilot-instructions.md with new paths

All suites now follow same pattern:
- tests/detection/ (181 tests)
- tests/nodejs-production/ (83 tests)
- tests/validation/ (165 tests)
