+++ .github/workflows/test-detection.yml
@@ -1,27 +1,30 @@
-name: Test Skills
+name: Detection Tests
 
 on:
   push:
     branches: [main]
     paths:
-      - 'plugin/skills/**'
-      - 'tests/**'
+      - 'plugin/skills/azure-deploy/**'
+      - 'tests/detection/**'
   pull_request:
     branches: [main]
     paths:
-      - 'plugin/skills/**'
-      - 'tests/**'
+      - 'plugin/skills/azure-deploy/**'
+      - 'tests/detection/**'
   workflow_dispatch:
 
 jobs:
   test:
-    name: Test (Node ${{ matrix.node-version }})
     runs-on: ubuntu-latest
     
     strategy:
       matrix:
         node-version: [20, 22]
     
+    defaults:
+      run:
+        working-directory: tests/detection
+
     steps:
       - name: Checkout repository
         uses: actions/checkout@v4
@@ -31,27 +34,35 @@ jobs:
         with:
           node-version: ${{ matrix.node-version }}
           cache: 'npm'
-          cache-dependency-path: tests/package-lock.json
+          cache-dependency-path: tests/detection/package-lock.json
 
       - name: Install dependencies
-        working-directory: tests
-        run: npm ci
+        run: npm ci || npm install
 
-      - name: Run detection tests
-        working-directory: tests
-        run: npm test -- --coverage --ci
+      - name: Run tests
+        run: npm test
 
-      - name: Upload coverage report
-        if: matrix.node-version == 22
+      - name: Run tests with coverage
+        run: npm run test:coverage
+
+      - name: Upload coverage to Codecov
+        if: matrix.node-version == 20
         uses: codecov/codecov-action@v4
         with:
-          directory: tests/coverage
-          flags: skill-detection
+          directory: tests/detection/coverage
+          flags: detection
           fail_ci_if_error: false
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
 
   lint:
     name: Lint
     runs-on: ubuntu-latest
+    
+    defaults:
+      run:
+        working-directory: tests/detection
+
     steps:
       - uses: actions/checkout@v4
       
@@ -60,14 +71,12 @@ jobs:
         with:
           node-version: 20
           cache: 'npm'
-          cache-dependency-path: tests/package-lock.json
+          cache-dependency-path: tests/detection/package-lock.json
 
       - name: Install dependencies
-        working-directory: tests
-        run: npm ci
+        run: npm ci || npm install
 
       - name: Check fixture/expectation sync
-        working-directory: tests
         run: |
           node -e "
             const { validateFixturesHaveExpectations } = require('./utils/fixture-loader');
