+++ tests/detection/__tests__/multiService.test.js
@@ -5,8 +5,8 @@
  * that should use Azure Developer CLI with Infrastructure as Code.
  */
 
-const { detectAppType } = require('../../src/detection/appTypeDetector');
-const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../../src/detection/serviceMapping');
+const { detectAppType } = require('../src/appTypeDetector');
+const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../src/serviceMapping');
 
 describe('Multi-Service Architecture Detection', () => {
   describe('Monorepo Detection', () => {
