+++ tests/detection/__tests__/nodejs.test.js
@@ -5,8 +5,8 @@
  * and routing to appropriate Azure services.
  */
 
-const { detectAppType } = require('../../src/detection/appTypeDetector');
-const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../../src/detection/serviceMapping');
+const { detectAppType } = require('../src/appTypeDetector');
+const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../src/serviceMapping');
 
 describe('Node.js Framework Detection', () => {
   describe('Next.js', () => {
