+++ tests/detection/__tests__/python.test.js
@@ -4,8 +4,8 @@
  * Tests for detecting Python frameworks and routing to appropriate Azure services.
  */
 
-const { detectAppType } = require('../../src/detection/appTypeDetector');
-const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../../src/detection/serviceMapping');
+const { detectAppType } = require('../src/appTypeDetector');
+const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../src/serviceMapping');
 
 describe('Python Framework Detection', () => {
   describe('Azure Functions', () => {
