+++ tests/detection/__tests__/static.test.js
@@ -4,8 +4,8 @@
  * Tests for detecting pure static HTML sites.
  */
 
-const { detectAppType } = require('../../src/detection/appTypeDetector');
-const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../../src/detection/serviceMapping');
+const { detectAppType } = require('../src/appTypeDetector');
+const { AZURE_SERVICES, SKILL_ROUTES, CONFIDENCE_LEVELS } = require('../src/serviceMapping');
 
 describe('Static Site Detection', () => {
   describe('Pure HTML Sites', () => {
