+++ tests/detection/fixtures/projects/angular/angular.json
