+++ tests/detection/fixtures/projects/angular/package.json
