+++ tests/detection/fixtures/projects/astro/astro.config.mjs
