+++ tests/detection/fixtures/projects/astro/package.json
