+++ tests/detection/fixtures/projects/azd-fullstack/api/package.json
