+++ tests/detection/fixtures/projects/azd-fullstack/azure.yaml
