+++ tests/detection/fixtures/projects/azd-fullstack/frontend/package.json
