+++ tests/detection/fixtures/projects/azd-fullstack/infra/main.bicep
