+++ tests/detection/fixtures/projects/django-app/requirements.txt
