+++ tests/detection/fixtures/projects/docker-compose-multi/docker-compose.yml
