+++ tests/detection/fixtures/projects/docker-node/Dockerfile
