+++ tests/detection/fixtures/projects/docker-node/package.json
