+++ tests/detection/fixtures/projects/dotnet-webapp/WebApp.csproj
