+++ tests/detection/fixtures/projects/express-api/package.json
