+++ tests/detection/fixtures/projects/fastapi-app/requirements.txt
