+++ tests/detection/fixtures/projects/functions-dotnet/FunctionsApp.csproj
