+++ tests/detection/fixtures/projects/functions-node/hello/function.json
