+++ tests/detection/fixtures/projects/functions-node/host.json
