+++ tests/detection/fixtures/projects/functions-node/package.json
