+++ tests/detection/fixtures/projects/functions-python/function_app.py
