+++ tests/detection/fixtures/projects/functions-python/requirements.txt
