+++ tests/detection/fixtures/projects/gatsby/gatsby-config.js
