+++ tests/detection/fixtures/projects/gatsby/package.json
