+++ tests/detection/fixtures/projects/nextjs-ssr/next.config.js
