+++ tests/detection/fixtures/projects/nextjs-ssr/package.json
