+++ tests/detection/fixtures/projects/nextjs-static/next.config.js
