+++ tests/detection/fixtures/projects/nextjs-static/package.json
