+++ tests/detection/fixtures/projects/react-vite/package.json
