+++ tests/detection/fixtures/projects/react-vite/vite.config.js
