+++ tests/detection/fixtures/projects/spring-boot/pom.xml
