+++ tests/detection/fixtures/projects/vue-vite/package.json
