+++ tests/detection/fixtures/projects/vue-vite/vite.config.js
