+++ tests/detection/jest.config.js
