+++ tests/detection/package.json
