+++ tests/detection/src/appTypeDetector.js
