+++ tests/detection/src/filePatterns.js
