+++ tests/detection/utils/fixture-loader.js
